Use uint16_t for rdataset record counts
Record counts are limited to 16 bits by the DNS wire and internal
rdataset formats. Reflect that limit in the API types and use the shared
16-bit accessors for negative-cache data.

+++ lib/dns/include/dns/ncache.h
@@ -94,8 +94,7 @@ dns_ncache_add(dns_message_t *message, dns_db_t *cache, dns_dbnode_t *node,
 
 isc_result_t
 dns_ncache_towire(dns_rdataset_t *rdataset, dns_compress_t *cctx,
-		  isc_buffer_t *target, unsigned int options,
-		  unsigned int *countp);
+		  isc_buffer_t *target, unsigned int options, uint16_t *countp);
 /*%<
  * Convert the negative caching rdataset 'rdataset' to wire format,
  * compressing names as specified in 'cctx', and storing the result in
