+++ lib/dns/include/dns/rdatalist.h
@@ -124,7 +124,7 @@ dns__rdatalist_current(dns_rdataset_t *rdataset, dns_rdata_t *rdata);
 void
 dns__rdatalist_clone(const dns_rdataset_t  *source,
 		     dns_rdataset_t *target DNS__DB_FLARG);
-unsigned int
+uint16_t
 dns__rdatalist_count(dns_rdataset_t *rdataset);
 isc_result_t
 dns__rdatalist_addnoqname(dns_rdataset_t *rdataset, dns_name_t *name);
